Move string extension into new tools-common project
