+++ kt/api-generator/src/main/kotlin/godot/codegen/models/enriched/EnrichedClass.kt
@@ -1,6 +1,6 @@
 package godot.codegen.models.enriched
 
-import godot.codegen.extensions.escapeUnderscore
+import godot.tools.common.extensions.escapeUnderscore
 import godot.codegen.models.Class
 import godot.codegen.models.custom.AdditionalImport
 import godot.codegen.traits.TypedTrait
