+++ kt/api-generator/src/main/kotlin/godot/codegen/models/enriched/EnrichedMethod.kt
@@ -1,7 +1,7 @@
 package godot.codegen.models.enriched
 
 import com.squareup.kotlinpoet.ANY
-import godot.codegen.extensions.convertToCamelCase
+import godot.tools.common.extensions.convertToCamelCase
 import godot.codegen.extensions.getTypeClassName
 import godot.codegen.extensions.isObjectSubClass
 import godot.codegen.workarounds.sanitizeApiType
