+++ kt/api-generator/src/main/kotlin/godot/codegen/models/enriched/EnrichedProperty.kt
@@ -2,7 +2,7 @@ package godot.codegen.models.enriched
 
 import com.squareup.kotlinpoet.ANY
 import godot.tools.common.constants.GodotTypes
-import godot.codegen.extensions.convertToCamelCase
+import godot.tools.common.extensions.convertToCamelCase
 import godot.codegen.extensions.getTypeClassName
 import godot.codegen.extensions.isObjectSubClass
 import godot.codegen.workarounds.sanitizeApiType
